upc_access.c (__getstf2, __getsxf2): Define temporary as "long double" rather than "double" to avoid loss of precision.
libupc/
	* libupc/smp/upc_access.c (__getstf2, __getsxf2):
	Define temporary as "long double" rather than "double" to avoid
	loss of precision.
	* testsuite/libupc.upc/intrepid/test29.upc: Improve the test of
	"long double" by adding checks that required accuracy
	exceeds the accuracy of "double".

From-SVN: r178994

+++ libupc/smp/upc_access.c
@@ -310,7 +310,7 @@ long double
 __getstf2 (upc_shared_ptr_t p)
 {
   const long double *addr = (long double *) __upc_access_sptr_to_addr (p);
-  double result;
+  long double result;
   GUPCR_FENCE ();
   result = *addr;
   GUPCR_READ_FENCE ();
@@ -322,7 +322,7 @@ long double
 __getsxf2 (upc_shared_ptr_t p)
 {
   const long double *addr = (long double *) __upc_access_sptr_to_addr (p);
-  double result;
+  long double result;
   GUPCR_FENCE ();
   result = *addr;
   GUPCR_READ_FENCE ();
